Fix hours max\min type

See merge request irlix/value-estimation-service/backend!25

+++ src/main/java/ru/irlix/evaluation/dto/response/TaskResponse.java
@@ -21,9 +21,9 @@ public class TaskResponse {
 
     private Integer repeatCount;
 
-    private Integer hoursMin;
+    private Double hoursMin;
 
-    private Integer hoursMax;
+    private Double hoursMax;
 
     private Long roleId;
